requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
pillow>=10.0
matplotlib>=3.7
pandas>=2.0
tqdm>=4.65
nltk>=3.8
gradio>=4.0
huggingface_hub>=0.20
streamlit>=1.30
